+++ user_events.py
@@ -33,6 +33,14 @@ from werkzeug.utils import secure_filename
 from glob import glob
 from os import remove, path, getcwd, makedirs
 
+import logging
+from time import gmtime
+
+logging.Formatter.converter = gmtime
+logging.basicConfig(level=logging.INFO, datefmt='%Y-%m-%dT%H:%M:%S',
+                    format='%(asctime)-15s.%(msecs)03dZ %(levelname)-7s [%(threadName)-10s] : %(name)s - %(message)s')
+__logger = logging.getLogger("user_events.py")
+
 userbp = Blueprint('user_events', __name__, url_prefix=Config.URL_PREFIX+'/user-events')
 
 @userbp.route('/', methods=['GET', 'POST'])
@@ -184,6 +192,14 @@ def user_an_event(id):
     #             targetAudience_edit_list += [item.capitalize()]
     #     post['targetAudience'] = targetAudience_edit_list
     post['longDescription'] = post['longDescription'].replace("\n", "<br>")
+    if post['subEvents']:
+        for subEvent in post['subEvents']:
+            event = find_user_event(clickable_utility(subEvent['id']))
+            if event['eventStatus'] == 'approved':
+                subEvent['isPublished'] = True
+            else:
+                subEvent['isPublished'] = False
+
     return render_template("events/event.html", post=post, eventTypeMap=eventTypeMap,
                            isUser=True, apiKey=current_app.config['GOOGLE_MAP_VIEW_KEY'],
                            timestamp=datetime.now().timestamp(),
@@ -231,16 +247,16 @@ def user_an_event_edit(id):
 
                 success = s3_delete_reupload(id, post_by_id.get('platformEventId'),image_record.get("_id"))
                 if success:
-                    print("{}, s3: s3_delete_reupload()".format(image_record.get('status')))
+                    __logger.info("{}, s3: s3_delete_reupload()".format(image_record.get('status')))
                     updateResult = update_one(current_app.config['IMAGE_COLLECTION'],
                                                  condition={'eventId': id},
                                                  update={"$set": {'status': 'replaced',
                                                                   'eventId': id}}, upsert=True)
                     post_by_id['imageURL'] = current_app.config['ROKWIRE_IMAGE_LINK_FORMAT'].format(post_by_id.get('platformEventId'), image_record.get("_id"))
                     if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                        print("Failed to mark image record as replaced of event: {} in event edit page".format(id))
+                        __logger.error("Failed to mark image record as replaced of event: {} in event edit page".format(id))
                 else:
-                    print("reuploading image for event:{} failed in event edit page".format(id))
+                    __logger.error("reuploading image for event:{} failed in event edit page".format(id))
             elif get_user_event_status(id) == "approved":
                 if image_record and image_record.get('status') == 'deleted':
                     updateResult = update_one(current_app.config['IMAGE_COLLECTION'],
@@ -248,7 +264,7 @@ def user_an_event_edit(id):
                                               update={"$set": {'status': 'new',
                                                                'eventId': id}}, upsert=True)
                     if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                        print("Failed to mark image record as new of event: {} in event edit page".format(
+                        __logger.error("Failed to mark image record as new of event: {} in event edit page".format(
                             id))
                 else:
                     insertResult = insert_one(current_app.config['IMAGE_COLLECTION'], document={
@@ -257,15 +273,15 @@ def user_an_event_edit(id):
                     })
                     image_record = find_one(Config.IMAGE_COLLECTION, condition={"eventId": id})
                     if not insertResult.inserted_id:
-                        print("Failed to mark image record as new of event: {} in event edit page".format(id))
+                        __logger.error("Failed to mark image record as new of event: {} in event edit page".format(id))
                 file.save(
                     path.join(Config.WEBTOOL_IMAGE_MOUNT_POINT, id + '.' + filename.rsplit('.', 1)[1].lower()))
                 success = s3_image_upload(id, post_by_id.get("platformEventId"), image_record.get("_id"))
                 if success:
-                    print("{}, s3: s3_image_upload()".format(image_record.get('status')))
+                    __logger.info("{}, s3: s3_image_upload()".format(image_record.get('status')))
                     post_by_id['imageURL'] = current_app.config['ROKWIRE_IMAGE_LINK_FORMAT'].format(id, image_record.get("_id"))
                 else:
-                    print("initial image upload for event:{} failed in event edit page".format(id))
+                    __logger.error("initial image upload for event:{} failed in event edit page".format(id))
             elif file and '.' in filename and filename.rsplit('.', 1)[1].lower() in Config.ALLOWED_IMAGE_EXTENSIONS:
                 file.save(
                     path.join(Config.WEBTOOL_IMAGE_MOUNT_POINT, id + '.' + filename.rsplit('.', 1)[1].lower()))
@@ -275,21 +291,21 @@ def user_an_event_edit(id):
             if image_record:
                 success = s3_image_delete(id, post_by_id.get("platformEventId"), image_record.get("_id"))
                 if success:
-                    print("{}, s3: s3_delete_reupload()".format(image_record.get('status')))
+                    __logger.info("{}, s3: s3_delete_reupload()".format(image_record.get('status')))
                     updateResult = update_one(current_app.config['IMAGE_COLLECTION'],
                                           condition={'eventId': id},
                                           update={"$set": {'status': 'deleted',
                                                            'eventId': id}}, upsert=True)
                     post_by_id['imageURL'] = ''
                     if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                        print("Failed to mark image record as deleted of event: {} in event edit page".format(id))
+                        __logger.error("Failed to mark image record as deleted of event: {} in event edit page".format(id))
                 else:
-                    print("deleting image for event:{} on s3 failed in event edit page".format(id))
+                    __logger.error("deleting image for event:{} on s3 failed in event edit page".format(id))
             else:
                 try:
                     remove(glob(path.join(Config.WEBTOOL_IMAGE_MOUNT_POINT, id + '*'))[0])
                 except OSError:
-                    print("delete event:{} image failed in ".format(id))
+                    __logger.error("delete event:{} image failed in ".format(id))
         all_day_event = False
         if 'allDay' in request.form and request.form.get('allDay') == 'on':
             post_by_id['allDay'] = True
@@ -300,6 +316,10 @@ def user_an_event_edit(id):
             post_by_id['isEventFree'] = True
         else:
             post_by_id['isEventFree'] = False
+        if 'displayOnlyWithSuperEvent' in request.form and request.form.get('displayOnlyWithSuperEvent') == 'on':
+            post_by_id['displayOnlyWithSuperEvent'] = True
+        else:
+            post_by_id['displayOnlyWithSuperEvent'] = False
 
         if 'isVirtual' in request.form and request.form.get('isVirtual') == 'on':
             post_by_id['isVirtual'] = True
@@ -378,10 +398,18 @@ def user_an_event_edit(id):
             for old_sub_event in old_sub_events:
                 if new_sub_events is None or old_sub_event not in new_sub_events:
                     update_super_event_id(old_sub_event['id'], '')
+
         if new_sub_events is not None:
+            removed_list = list()
             for new_sub_event in new_sub_events:
-                if old_sub_events is None or new_sub_event not in old_sub_events:
-                    update_super_event_id(new_sub_event['id'], id)
+                try:
+                    if old_sub_events is None or new_sub_event not in old_sub_events:
+                        update_super_event_id(new_sub_event['id'], id)
+                except Exception as ex:
+                    removed_list.append(new_sub_event)
+                    pass
+        for deleted_sub_event in removed_list:
+            new_sub_events.remove(deleted_sub_event)
 
         old_title = find_one(current_app.config['EVENT_COLLECTION'],
                                   condition={"_id": ObjectId(id)})['title']
@@ -399,8 +427,9 @@ def user_an_event_edit(id):
                                                       "$set": {"subEvents": sub_event_list}
                                                   })
                         if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                            print("Failed to update the title of sub-event {} in super event {}".format(id, post_by_id['superEventID']))
+                            __logger.error("Failed to update the title of sub-event {} in super event {}".format(id, post_by_id['superEventID']))
             except Exception as ex:
+                __logger.exception(ex)
                 pass
 
         if 'timezone' in request.form:
@@ -490,10 +519,10 @@ def user_an_event_approve(id):
                                       update={"$set": {'status': 'new',
                                                        'eventId': id}}, upsert=True)
             if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                print("Failed to mark image record as new of event: {} upon event publishing".format(id))
+                __logger.error("Failed to mark image record as new of event: {} upon event publishing".format(id))
             approve_user_event(id)
-    except Exception:
-        traceback.print_exc()
+    except Exception as ex:
+        __logger.exception(ex)
     if success:
         return "success", 200
     else:
@@ -504,8 +533,8 @@ def user_an_event_approve(id):
 def user_an_event_disapprove(id):
     try:
         disapprove_user_event(id)
-    except Exception:
-        traceback.print_exc()
+    except Exception as ex:
+        __logger.exception(ex)
 
     return "success", 200
 
@@ -547,6 +576,8 @@ def add_new_event():
         new_event['isGroupPrivate'] = False
         if new_event.get('isEventFree') == 'on':
             new_event['isEventFree'] = True
+        if new_event.get('displayOnlyWithSuperEvent') == 'on':
+            new_event['displayOnlyWithSuperEvent'] = True
         new_event_id = create_new_user_event(new_event)
         if new_event['subEvents'] is not None:
             for subEvent in new_event['subEvents']:
@@ -586,7 +617,7 @@ def notification_event(id):
     message = request.form.get('message')
     data = {"type": "event_detail", "event_id": id}
     tokens = request.form.get('tokens').split(",")
-    print("notification: event platform id: %s , title: %s, message body: %s" % (id, title, message))
+    __logger.info("notification: event platform id: %s , title: %s, message body: %s" % (id, title, message))
     # send notification
     notification.send_notification(title, message, data, tokens)
     return "", 200
@@ -601,7 +632,7 @@ def get_devicetokens(id):
 @role_required("user")
 def userevent_delete(id):
     userEvent = find_user_event(id)
-    print("delete user event id: %s" % id)
+    __logger.info("delete user event id: %s" % id)
     sub_events = find_one(current_app.config['EVENT_COLLECTION'], condition={"_id": ObjectId(id)}).get('subEvents')
     if sub_events is not None:
         for sub_event in sub_events:
@@ -627,7 +658,7 @@ def userevent_delete(id):
                     if get_user_event_status(super_event['_id']) == "approved":
                         success = put_user_event(super_event['_id'])
                         if not success:
-                            print("updating super event in building block failed")
+                            __logger.error("updating super event in building block failed")
                     find = True
                     break
     delete_user_event(id)
@@ -635,20 +666,20 @@ def userevent_delete(id):
         try:
             remove(glob(path.join(Config.WEBTOOL_IMAGE_MOUNT_POINT, id + '*'))[0])
         except OSError:
-            print("delete event:{} image failed".format(id))
+            __logger.error("delete event:{} image failed".format(id))
     record = find_one(Config.IMAGE_COLLECTION, condition={"eventId": id})
     if record:
         success = s3_image_delete(id, userEvent.get("platformEventId"), record.get("_id"))
         if success:
-            print("{}, s3: s3_image_delete()".format(record.get('status')))
+            __logger.info("{}, s3: s3_image_delete()".format(record.get('status')))
             updateResult = update_one(current_app.config['IMAGE_COLLECTION'],
                                       condition={'eventId': id},
                                       update={"$set": {'status': 'deleted',
                                                        'eventId': id}}, upsert=True)
             if updateResult.modified_count == 0 and updateResult.matched_count == 0 and updateResult.upserted_id is None:
-                print("Failed to mark image record as deleted of event: {} in the deletion of event".format(id))
+                __logger.error("Failed to mark image record as deleted of event: {} in the deletion of event".format(id))
         else:
-            print("deleting image for event:{} failed in event deletion".format(id))
+            __logger.error("deleting image for event:{} failed in event deletion".format(id))
     return "", 200
 
 @userbp.route('/search', methods=['GET', 'POST'])
@@ -666,7 +697,7 @@ def search():
 def searchsub():
     if request.method == "GET":
        search_term = request.values.get("data")
-       return jsonify(beta_search(search_term))
+       return jsonify(group_subevents_search(search_term, get_admin_group_ids()))
     else:
        return jsonify([]), 200
 
@@ -679,7 +710,7 @@ def view_image(id):
         success = s3_image_download(id, event.get("platformEventId"), record.get("_id"))
         if success:
             try:
-                print("{}, s3: s3_image_download()".format(record.get('status')))
+                __logger.info("{}, s3: s3_image_download()".format(record.get('status')))
                 path_to_tmp_image = os.path.join(os.getcwd(), 'temp', id + ".jpg")
 
                 def get_image():
@@ -689,9 +720,9 @@ def view_image(id):
 
                 response = current_app.response_class(get_image(), mimetype='image/jpg')
                 return response
-            except Exception:
-                traceback.print_exc()
-                print("returning image for event:{} on s3 to user failed".format(id))
+            except Exception as ex:
+                __logger.exception(ex)
+                __logger.error("returning image for event:{} on s3 to user failed".format(id))
         else:
             abort(404)
     else:
@@ -709,7 +740,7 @@ def sub_event(platformEventId):
         eventId = clickable_utility(platformEventId)
         return redirect(url_for('user_events.user_an_event', id=eventId))
 
-    except Exception:
-        traceback.print_exc()
-        print("Redirect for platformEventId {} failed".format(platformEventId))
+    except Exception as ex:
+        __logger.exception(ex)
+        __logger.error("Redirect for platformEventId {} failed".format(platformEventId))
         abort(500)